Class/Option: Options on Class and FindClasses()
Update classers to use a sync.Map
Updated classers cache changing get() to find() to check a Classer

Update get() to always return a classer or unknownClasser instance along with a function to remote the instance if it was allocated.

Add FindChileClasses() to return all the Class objects associated with a Classer type.

Add Options Help(), Type(), Format()